requirements.txt
asgiref==3.5.0
cmake==3.22.2
crispy-tailwind==0.5.0
Django==4.0.3
django-crispy-forms==1.14.0
numpy==1.22.3
opencv-contrib-python==4.5.5.64
Pillow==9.0.1
pytz==2021.3
sqlparse==0.4.2
